refactor: inline CharLengthMixin
